+++ lib/folder_file_saver.dart
@@ -10,13 +10,6 @@ class FolderFileSaver {
   /// return 0 permission is PERMISSION_GRANTED
   /// return 1 permission is PERMISSION_IS_DENIED
   /// return 2 permission is PERMISSION_IS_DENIED (Don't Ask Again)
-  static Future<int> checkPermission() {
-    return _channel.invokeMethod('checkPermission');
-  }
-
-  /// return 0 permission is PERMISSION_GRANTED
-  /// return 1 permission is PERMISSION_IS_DENIED
-  /// return 2 permission is PERMISSION_IS_DENIED (Don't Ask Again)
   static Future<int> requestPermission() async {
     await _channel.invokeMethod('requestPermission');
     final completer = new Completer<int>();